ESIMW-1001: Allow matching of local part of email address in alias.
I'm not sure this is a good idea, but it's one way to work around the appdeploy plugin not handling `@` signs well. If there is not a property containing the full alias of the email cert address, it then checks if there is an alias containing the local part of the email cert address.